+++ packages/synapse-interface/constants/tokens/bridgeable.ts
@@ -17,6 +17,7 @@ import hyperjumpLogo from '@assets/icons/hyperjump.svg'
 import jewelLogo from '@assets/icons/jewel.svg'
 import klayLogo from '@assets/icons/klay.svg'
 import l2daoLogo from '@assets/icons/l2dao.svg'
+import lusdLogo from '@assets/icons/lusd.svg'
 import maticLogo from '@assets/icons/matic.svg'
 import movrLogo from '@assets/icons/movr.svg'
 import nethLogo from '@assets/icons/neth.svg'
@@ -42,6 +43,18 @@ import wethLogo from '@assets/icons/weth.svg'
 import { Token } from '@/utils/types'
 import * as CHAINS from '@/constants/chains/master'
 
+// Priority ranks:
+// 100: chain's major stablecoins (native DAI, USDC, USDT)
+// 125: chain's major stablecoins (bridged)
+// 150: ETH (WETH if chain's native asset is ETH)
+// 200: rest of the chain's stablecoins
+// 250: SYN, biggest partner tokens (GMX, JEWEL, etc)
+// 300: chain's native asset (AVAX, FTM, MATIC, KLAY, etc)
+// 350: wrapped versions of native asset (WAVAX, WFTM, etc)
+// 400: synAssets
+// 500: nAssets
+// 600: everything else
+
 export const GOHM = new Token({
   addresses: {
     [CHAINS.ETH.id]: '0x0ab87046fBb341D058F17CBC4c1133F25a20a52f',
@@ -65,7 +78,7 @@ export const GOHM = new Token({
   swapableType: 'OHM',
   color: 'gray',
   visibilityRank: 40,
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'gOHM',
 })
 
@@ -80,7 +93,7 @@ export const LINK = new Token({
   logo: linkLogo,
   swapableType: 'LINK',
   color: 'blue',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'LINK',
 })
 
@@ -95,7 +108,7 @@ export const HIGH = new Token({
   logo: highLogo,
   swapableType: 'HIGH',
   color: 'cyan',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'HIGH',
 })
 
@@ -112,7 +125,7 @@ export const JUMP = new Token({
   docUrl: '',
   swapableType: 'JUMP',
   color: 'cyan',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'JUMP',
 })
 
@@ -128,7 +141,7 @@ export const SFI = new Token({
   docUrl: '',
   swapableType: 'SFI',
   color: 'red',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'SFI',
 })
 
@@ -145,7 +158,7 @@ export const DOG = new Token({
   docUrl: '',
   swapableType: 'DOG',
   color: 'yellow',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'DOG',
 })
 
@@ -163,7 +176,7 @@ export const NFD = new Token({
   docUrl: '',
   swapableType: 'NFD',
   color: 'yellow',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'NFD',
 })
 
@@ -179,7 +192,7 @@ export const SOLAR = new Token({
   docUrl: '',
   swapableType: 'SOLAR',
   color: 'orange',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'veSOLAR',
 })
 
@@ -197,7 +210,7 @@ export const GMX = new Token({
   logo: gmxLogo,
   docUrl: '',
   swapableType: 'GMX',
-  priorityRank: 6,
+  priorityRank: 250,
   color: 'blue',
   routeSymbol: 'GMX',
 })
@@ -217,7 +230,7 @@ export const SDT = new Token({
   docUrl: '',
   swapableType: 'SDT',
   color: 'gray',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'SDT',
 })
 
@@ -234,7 +247,7 @@ export const NEWO = new Token({
   docUrl: '',
   swapableType: 'NEWO',
   color: 'yellow',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'NEWO',
 })
 
@@ -248,7 +261,7 @@ export const PEPE = new Token({
   name: 'Pepe',
   logo: pepeLogo,
   swapableType: 'PEPE',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'PEPE',
   color: 'green',
 })
@@ -265,7 +278,7 @@ export const VSTA = new Token({
   docUrl: '',
   swapableType: 'VSTA',
   color: 'gray',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'VSTA',
 })
 
@@ -287,7 +300,7 @@ export const H2O = new Token({
   docUrl: '',
   swapableType: 'H2O',
   color: 'cyan',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'H2O',
 })
 
@@ -303,7 +316,7 @@ export const L2DAO = new Token({
   docUrl: '',
   swapableType: 'L2DAO',
   color: 'cyan',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'L2DAO',
 })
 
@@ -319,7 +332,7 @@ export const PLS = new Token({
   docUrl: '',
   swapableType: 'PLS',
   color: 'green',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'PLS',
 })
 
@@ -336,7 +349,7 @@ export const AGEUR = new Token({
   docUrl: '',
   swapableType: 'AGEUR',
   color: 'yellow',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'agEUR',
 })
 
@@ -355,7 +368,7 @@ export const UNIDX = new Token({
   docUrl: '',
   swapableType: 'UNIDX',
   color: 'gray',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'UNIDX',
 })
 
@@ -371,7 +384,7 @@ export const BUSD = new Token({
   swapableType: 'BUSD',
   swapableOn: [CHAINS.BNB.id],
   color: 'yellow',
-  priorityRank: 2,
+  priorityRank: 200,
   routeSymbol: 'BUSD',
 })
 
@@ -430,7 +443,7 @@ export const USDC = new Token({
     CHAINS.CANTO.id,
   ],
   color: 'blue',
-  priorityRank: 1,
+  priorityRank: 100,
   routeSymbol: 'USDC',
 })
 
@@ -448,7 +461,7 @@ export const METISUSDC = new Token({
   swapableType: 'USD',
   swapableOn: [CHAINS.METIS.id],
   color: 'blue',
-  priorityRank: 3,
+  priorityRank: 125,
   routeSymbol: 'm.USDC',
 })
 
@@ -498,7 +511,7 @@ export const USDT = new Token({
     CHAINS.CANTO.id,
   ],
   visibilityRank: 100,
-  priorityRank: 1,
+  priorityRank: 100,
   routeSymbol: 'USDT',
 })
 
@@ -531,7 +544,7 @@ export const DAI = new Token({
   ],
   color: 'yellow',
   visibilityRank: 100,
-  priorityRank: 1,
+  priorityRank: 100,
   routeSymbol: 'DAI',
 })
 
@@ -551,7 +564,7 @@ export const WBTC = new Token({
   logo: wbtcLogo,
   swapableType: 'WBTC',
   color: 'orange',
-  priorityRank: 3,
+  priorityRank: 300,
   routeSymbol: 'WBTC',
 })
 
@@ -566,7 +579,7 @@ export const WETHE = new Token({
   swapableType: 'ETH',
   swapableOn: [CHAINS.AVALANCHE.id],
   color: 'sky',
-  priorityRank: 2,
+  priorityRank: 150,
   routeSymbol: 'WETH.e',
 })
 
@@ -581,7 +594,7 @@ export const ONEETH = new Token({
   swapableType: 'ETH',
   swapableOn: [CHAINS.HARMONY.id],
   color: 'sky',
-  priorityRank: 2,
+  priorityRank: 600,
   routeSymbol: '1ETH',
 })
 
@@ -612,7 +625,7 @@ export const SYN = new Token({
   swapableType: 'SYN',
   color: 'purple',
   visibilityRank: 90,
-  priorityRank: 2,
+  priorityRank: 250,
   routeSymbol: 'SYN',
 })
 
@@ -628,7 +641,7 @@ export const FRAX = new Token({
   logo: fraxLogo,
   swapableType: 'FRAX',
   color: 'gray',
-  priorityRank: 6,
+  priorityRank: 200,
   routeSymbol: 'FRAX',
 })
 
@@ -645,7 +658,7 @@ export const SYNFRAX = new Token({
   logo: synapseLogo,
   swapableType: 'FRAX',
   color: 'gray',
-  priorityRank: 4,
+  priorityRank: 400,
   routeSymbol: 'synFRAX',
 })
 
@@ -686,7 +699,7 @@ export const NUSD = new Token({
   ],
   color: 'purple',
   visibilityRank: 50,
-  priorityRank: 5,
+  priorityRank: 500,
   routeSymbol: 'nUSD',
 })
 
@@ -702,7 +715,7 @@ export const NOTE = new Token({
   swapableOn: [CHAINS.CANTO.id],
   color: 'green',
   visibilityRank: 90,
-  priorityRank: 3,
+  priorityRank: 100,
   routeSymbol: 'NOTE',
 })
 
@@ -735,7 +748,7 @@ export const NETH = new Token({
   ],
   color: 'purple',
   visibilityRank: 50,
-  priorityRank: 5,
+  priorityRank: 500,
   routeSymbol: 'nETH',
 })
 
@@ -757,7 +770,7 @@ export const ETH = new Token({
   swapableType: 'ETH',
   color: 'sky',
   visibilityRank: 101,
-  priorityRank: 2,
+  priorityRank: 150,
   swapableOn: [
     CHAINS.ARBITRUM.id,
     CHAINS.BASE.id,
@@ -779,7 +792,7 @@ export const MOVR = new Token({
   isNative: true,
   swapableType: 'MOVR',
   color: 'purple',
-  priorityRank: 3,
+  priorityRank: 300,
   routeSymbol: 'MOVR',
 })
 
@@ -799,7 +812,7 @@ export const AVAX = new Token({
   swapableType: 'AVAX',
   color: 'red',
   visibilityRank: 90,
-  priorityRank: 3,
+  priorityRank: 300,
   routeSymbol: 'AVAX',
 })
 
@@ -813,7 +826,7 @@ export const WMOVR = new Token({
   logo: movrLogo,
   swapableType: 'MOVR',
   color: 'purple',
-  priorityRank: 3,
+  priorityRank: 350,
   routeSymbol: 'WMOVR',
 })
 
@@ -828,7 +841,7 @@ export const WAVAX = new Token({
   swapableType: 'AVAX',
   color: 'red',
   visibilityRank: 90,
-  priorityRank: 3,
+  priorityRank: 350,
   routeSymbol: 'WAVAX',
 })
 
@@ -846,7 +859,7 @@ export const JEWEL = new Token({
   color: 'lime',
   isNative: true,
   swapableType: 'JEWEL',
-  priorityRank: 3,
+  priorityRank: 250,
   routeSymbol: 'JEWEL',
 })
 
@@ -861,7 +874,7 @@ export const WJEWEL = new Token({
   swapableType: 'JEWEL',
   swapableOn: [CHAINS.HARMONY.id],
   color: 'lime',
-  priorityRank: 3,
+  priorityRank: 350,
   routeSymbol: 'WJEWEL',
 })
 
@@ -876,7 +889,7 @@ export const SYNJEWEL = new Token({
   swapableType: 'JEWEL',
   swapableOn: [CHAINS.HARMONY.id],
   color: 'lime',
-  priorityRank: 4,
+  priorityRank: 400,
   routeSymbol: 'synJEWEL',
 })
 
@@ -891,7 +904,7 @@ export const XJEWEL = new Token({
   logo: jewelLogo,
   swapableType: 'XJEWEL',
   color: 'lime',
-  priorityRank: 3,
+  priorityRank: 350,
   routeSymbol: 'xJEWEL',
 })
 
@@ -910,7 +923,7 @@ export const USDCe = new Token({
   color: 'blue',
   swapableOn: [CHAINS.AVALANCHE.id, CHAINS.ARBITRUM.id, CHAINS.OPTIMISM.id],
   visibilityRank: 100,
-  priorityRank: 1,
+  priorityRank: 125,
   routeSymbol: 'USDC.e',
 })
 
@@ -927,7 +940,7 @@ export const USDTe = new Token({
   swapableOn: [CHAINS.AVALANCHE.id],
   visibilityRank: 100,
   color: 'green',
-  priorityRank: 1,
+  priorityRank: 125,
   routeSymbol: 'USDT.e',
 })
 
@@ -945,7 +958,7 @@ export const SUSD = new Token({
   swapableType: 'USD',
   swapableOn: [CHAINS.OPTIMISM.id],
   visibilityRank: 100,
-  priorityRank: 1,
+  priorityRank: 200,
   routeSymbol: 'sUSD',
 })
 
@@ -966,7 +979,7 @@ export const WSOHM = new Token({
   color: 'gray',
   swapableType: 'OHM',
   visibilityRank: 40,
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: 'wsOHM',
 })
 
@@ -982,7 +995,7 @@ export const ONEDAI = new Token({
   swapableOn: [CHAINS.HARMONY.id],
   color: 'yellow',
   visibilityRank: 100,
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: '1DAI',
 })
 
@@ -1000,7 +1013,7 @@ export const ONEUSDC = new Token({
   swapableType: 'USD',
   swapableOn: [CHAINS.HARMONY.id],
   color: 'blue',
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: '1USDC',
 })
 
@@ -1018,7 +1031,7 @@ export const ONEUSDT = new Token({
   swapableType: 'USD',
   swapableOn: [CHAINS.HARMONY.id],
   visibilityRank: 100,
-  priorityRank: 6,
+  priorityRank: 600,
   routeSymbol: '1USDT',
 })
 
@@ -1038,7 +1051,7 @@ export const BTCB = new Token({
   logo: btcLogo,
   swapableType: 'BTC.b',
   color: 'orange',
-  priorityRank: 3,
+  priorityRank: 300,
   routeSymbol: 'BTC.b',
 })
 
@@ -1054,7 +1067,7 @@ export const DAIE = new Token({
   swapableOn: [CHAINS.AVALANCHE.id],
   color: 'yellow',
   visibilityRank: 100,
-  priorityRank: 2,
+  priorityRank: 125,
   routeSymbol: 'DAI.e',
 })
 
@@ -1073,7 +1086,7 @@ export const KLAY = new Token({
   isNative: true,
   swapableType: 'KLAY',
   color: 'red',
-  priorityRank: 3,
+  priorityRank: 300,
   routeSymbol: 'KLAY',
 })
 
@@ -1089,7 +1102,7 @@ export const WKLAY = new Token({
   logo: klayLogo,
   swapableType: 'WKLAY',
   color: 'red',
-  priorityRank: 3,
+  priorityRank: 350,
   routeSymbol: 'WKLAY',
 })
 
@@ -1106,7 +1119,7 @@ export const MATIC = new Token({
   swapableType: 'MATIC',
   color: 'blue',
   visibilityRank: 90,
-  priorityRank: 3,
+  priorityRank: 300,
   routeSymbol: 'MATIC',
 })
 
@@ -1121,7 +1134,7 @@ export const WMATIC = new Token({
   swapableType: 'MATIC',
   color: 'blue',
   visibilityRank: 90,
-  priorityRank: 3,
+  priorityRank: 350,
   routeSymbol: 'WMATIC',
 })
 
@@ -1137,7 +1150,7 @@ export const FTM = new Token({
   swapableType: 'FTM',
   swapableOn: [CHAINS.FANTOM.id],
   color: 'blue',
-  priorityRank: 5,
+  priorityRank: 300,
   routeSymbol: 'FTM',
 })
 
@@ -1152,7 +1165,7 @@ export const WFTM = new Token({
   swapableType: 'FTM',
   swapableOn: [CHAINS.FANTOM.id],
   color: 'blue',
-  priorityRank: 5,
+  priorityRank: 350,
   routeSymbol: 'WFTM',
 })
 
@@ -1174,18 +1187,17 @@ export const WETH = new Token({
   logo: wethLogo,
   swapableType: 'ETH',
   color: 'sky',
-  priorityRank: 3,
+  priorityRank: 350,
   routeSymbol: 'WETH',
 })
 
-export const CRVUSDC = new Token({
+export const CRVUSD = new Token({
   visibilityRank: 101,
   addresses: {
+    [CHAINS.ETH.id]: '0xf939E0A03FB07F59A73314E73794Be0E57ac1b4E',
     [CHAINS.BASE.id]: '0x417Ac0e078398C154EdFadD9Ef675d30Be60Af93',
   },
-  decimals: {
-    [CHAINS.BASE.id]: 18,
-  },
+  decimals: 18,
   swapExceptions: {},
   symbol: 'crvUSD',
   name: 'Curve.fi USD',
@@ -1193,8 +1205,25 @@ export const CRVUSDC = new Token({
   swapableType: 'USD',
   swapableOn: [],
   color: 'yellow',
-  priorityRank: 1,
+  priorityRank: 200,
   routeSymbol: 'crvUSD',
+})
+
+export const LUSD = new Token({
+  visibilityRank: 101,
+  addresses: {
+    [CHAINS.ETH.id]: '0x5f98805A4E8be255a32880FDeC7F6728C6568bA0',
+  },
+  decimals: 18,
+  swapExceptions: {},
+  symbol: 'LUSD',
+  name: 'Liquity USD',
+  logo: lusdLogo,
+  swapableType: 'USD',
+  swapableOn: [],
+  color: 'blue',
+  priorityRank: 200,
+  routeSymbol: 'LUSD',
 })
 
 export const USDBC = new Token({
@@ -1212,6 +1241,6 @@ export const USDBC = new Token({
   swapableType: 'USD',
   swapableOn: [],
   color: 'blue',
-  priorityRank: 1,
+  priorityRank: 125,
   routeSymbol: 'USDbC',
 })